web_app/requirements.txt
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3
requests==2.31.0
beautifulsoup4==4.12.2
plotly==5.17.0
dart-fss==0.4.6
python-dotenv==1.0.0
uuid